+++ ZXingObjC/qrcode/decoder/ZXQRCodeVersion.h
@@ -62,7 +62,7 @@
  */
 @interface ZXQRCodeECB : NSObject
 
-@property (nonatomic, assign, readonly) int count;
+@property (nonatomic, assign, readonly) NSUInteger count;
 @property (nonatomic, assign, readonly) int dataCodewords;
 
 - (id)initWithCount:(int)count dataCodewords:(int)dataCodewords;
