+++ ZXingObjC/qrcode/detector/ZXQRCodeFinderPattern.h
@@ -23,7 +23,7 @@
  */
 @interface ZXQRCodeFinderPattern : ZXResultPoint
 
-@property (nonatomic, assign, readonly) int count;
+@property (nonatomic, assign, readonly) NSUInteger count;
 @property (nonatomic, assign, readonly) float estimatedModuleSize;
 
 - (id)initWithPosX:(float)posX posY:(float)posY estimatedModuleSize:(float)estimatedModuleSize;
